+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/style/layers/internal/LayerNode.kt
@@ -49,6 +49,7 @@ internal class LayerNode(
   }
 
   private fun detachSource() {
+    logD(TAG, "detaching Source: $this")
     sourceState?.detachFromLayer(layerId, map)
   }